Begin working out application for mutually-recursive types
This required a _lot_ of thinking! Type application to an alias in the current
recursive group was not expressible before. What we need to do is add extra
constructors which capture this. They can be unified (nominally) in the same way
as applications of abstract types.

The idea (not implemented yet) is then to add all bindings as `Mutual (also new)
in the alias environment when desugaring the types, and then they can be added
properly to the alias environment afterwards.