Add infrastructure for temporal metadata in TableHandle types
It compiles, but there's still some dodginess around. In particular, I'm
unsure whether "TemporalMetadata.unspecified" (which we allow to unify
with anything) is safe, or we'll have to do a proper reference-based
unification here...

Similarly, a few things are still dependent upon the design and
implementation of the TableLit extension (for example, in the TC, we'd
want to ensure that the MD is compatible with a TableHandle annotation).